Renamed utility file to require_goto_statements
